Rename from `snake_case` to `kebab-case`
