+++ brackets-prisma-db/src/storage-handlers/update-handlers/index.ts
@@ -1,6 +1,6 @@
 export * from './group';
 export * from './match';
-export * from './match_game';
+export * from './match-game';
 export * from './participant';
 export * from './round';
 export * from './stage';
