+++ brackets-prisma-db/src/transformers/enum/index.ts
@@ -1,6 +1,6 @@
-export * from './grand_final_type';
-export * from './match_result';
-export * from './match_status';
-export * from './round_robin_mode';
-export * from './seed_ordering';
-export * from './stage_type';
+export * from './grand-final-type';
+export * from './match-result';
+export * from './match-status';
+export * from './round-robin-mode';
+export * from './seed-ordering';
+export * from './stage-type';
