+++ brackets-prisma-db/src/transformers/model/index.ts
@@ -1,8 +1,8 @@
 export * from './group';
 export * from './match';
-export * from './match_game';
+export * from './match-game';
 export * from './participant';
-export * from './participant_result';
+export * from './participant-result';
 export * from './round';
 export * from './stage';
-export * from './stage_settings';
+export * from './stage-settings';
